refactor: 포트폴리오 get요청 리액트쿼리로 변경 및 컴포넌트 세분화
